requirements.txt
# QuantConnect Algorithm Dependencies
# Core QuantConnect libraries are provided by the platform

# Data Analysis and Numerical Computing
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0

# Technical Analysis
TA-Lib>=0.4.24
pandas-ta>=0.3.14b

# Configuration Management
PyYAML>=6.0
python-dotenv>=0.19.0

# Logging and Utilities
colorlog>=6.6.0
python-dateutil>=2.8.2

# Testing Framework
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.6.0

# Jupyter Notebooks (for analysis)
jupyter>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0

# Performance Analysis
pyfolio>=0.9.2
empyrical>=0.5.5

# Optimization
scipy>=1.7.0
scikit-learn>=1.0.0

# Development Tools
black>=22.0.0
flake8>=4.0.0
mypy>=0.910

# Documentation
sphinx>=4.0.0
sphinx-rtd-theme>=1.0.0
